Enhance User classes with code improvements

+++ src/Services/user/UserService.cs
@@ -28,6 +28,10 @@ namespace Backend_Teamwork.src.Services.user
         public async Task<List<UserReadDto>> GetAllAsync()
         {
             var UserList = await _userRepository.GetAllAsync();
+            if (UserList.Count == 0)
+            {
+                throw CustomException.NotFound($"Users not found");
+            }
             return _mapper.Map<List<User>, List<UserReadDto>>(UserList);
         }
 
@@ -40,6 +44,10 @@ namespace Backend_Teamwork.src.Services.user
         // Retrieves users with pagination options
         public async Task<List<UserReadDto>> GetUsersByPage(PaginationOptions paginationOptions)
         {
+            if (paginationOptions == null)
+            {
+                throw CustomException.BadRequest("pagination options cannot be null.");
+            }
             // Validate pagination options
             if (paginationOptions.Limit <= 0)
             {
@@ -51,6 +59,11 @@ namespace Backend_Teamwork.src.Services.user
                 throw CustomException.BadRequest("Offset should be 0 or greater.");
             }
             var UserList = await _userRepository.GetAllAsync(paginationOptions);
+
+            if (UserList.Count == 0)
+            {
+                throw CustomException.NotFound("No users found.");
+            }
             return _mapper.Map<List<User>, List<UserReadDto>>(UserList);
         }
 
@@ -77,6 +90,10 @@ namespace Backend_Teamwork.src.Services.user
             }
 
             var UserCreated = await _userRepository.CreateOneAsync(user);
+            if (UserCreated == null)
+            {
+                throw CustomException.BadRequest("Failed to create user.");
+            }
             return _mapper.Map<User, UserReadDto>(UserCreated);
         }
 
@@ -86,7 +103,7 @@ namespace Backend_Teamwork.src.Services.user
             var foundUser = await _userRepository.GetByIdAsync(id);
             if (foundUser == null)
             {
-                throw CustomException.NotFound("User not found.");
+                throw CustomException.NotFound($"User with id: {id} not found");
             }
             return _mapper.Map<User, UserReadDto>(foundUser);
         }
@@ -94,36 +111,65 @@ namespace Backend_Teamwork.src.Services.user
         // Deletes a user by their ID
         public async Task<bool> DeleteOneAsync(Guid id)
         {
+            if (id == Guid.Empty)
+            {
+                throw CustomException.BadRequest("Invalid user ID");
+            }
             var foundUser = await _userRepository.GetByIdAsync(id);
             if (foundUser == null)
             {
-                throw CustomException.NotFound("User not found.");
+                throw CustomException.NotFound($"User with ID {id} not found.");
             }
-            return await _userRepository.DeleteOneAsync(foundUser);
+            var DeletedUser = await _userRepository.DeleteOneAsync(foundUser);
+
+            // Check if the delete was successful
+            if (!DeletedUser)
+            {
+                throw CustomException.BadRequest("Failed to delete user.");
+            }
+
+            return DeletedUser;
         }
 
         // Updates a user by their ID
         public async Task<bool> UpdateOneAsync(Guid id, UserUpdateDto updateDto)
         {
+            if (id == Guid.Empty)
+            {
+                throw CustomException.BadRequest("Invalid user ID");
+            }
+
             var foundUser = await _userRepository.GetByIdAsync(id);
             if (foundUser == null)
             {
-                throw CustomException.NotFound("User not found.");
+                throw CustomException.NotFound($"User with ID {id} not found.");
             }
 
             // Map the update DTO to the existing User entity
             _mapper.Map(updateDto, foundUser);
 
-            return await _userRepository.UpdateOneAsync(foundUser);
+            var updatedUser = await _userRepository.UpdateOneAsync(foundUser);
+
+            // Check if the update was successful
+            if (!updatedUser)
+            {
+                throw CustomException.BadRequest("Failed to update user.");
+            }
+
+            return updatedUser;
         }
 
         // Retrieves a user by their email
         public async Task<UserReadDto> GetByEmailAsync(string email)
         {
+            if (string.IsNullOrWhiteSpace(email))
+            {
+                throw CustomException.BadRequest("Email is required");
+            }
             var user = await _userRepository.GetByEmailAsync(email);
             if (user == null)
             {
-                throw CustomException.NotFound("User not found.");
+                throw CustomException.NotFound($"User with email {email} not found.");
             }
             return _mapper.Map<User, UserReadDto>(user);
         }
@@ -131,32 +177,32 @@ namespace Backend_Teamwork.src.Services.user
         // Retrieves a user by their phone number
         public async Task<UserReadDto> GetByPhoneNumberAsync(string phoneNumber)
         {
+            if (string.IsNullOrWhiteSpace(phoneNumber))
+            {
+                throw CustomException.BadRequest("Phone Number is required");
+            }
+
             var user = await _userRepository.GetByPhoneNumberAsync(phoneNumber);
             if (user == null)
             {
                 throw CustomException.NotFound("User not found.");
             }
             return _mapper.Map<User, UserReadDto>(user);
-        }
-
-        // Retrieves a user by their name
-        public async Task<UserReadDto> GetByNameAsync(string name)
-        {
-            var user = await _userRepository.GetByNameAsync(name);
-            if (user == null)
-            {
-                throw CustomException.NotFound("User not found.");
-            }
-            return _mapper.Map<User, UserReadDto>(user);
+            ;
         }
 
         // Signs in a user with their credentials
         public async Task<string> SignInAsync(UserCreateDto createDto)
         {
+            if (createDto == null)
+            {
+                throw CustomException.BadRequest("User data cannot be null.");
+            }
+
             var foundUser = await _userRepository.GetByEmailAsync(createDto.Email);
             if (foundUser == null)
             {
-                throw CustomException.NotFound("User not found.");
+                throw CustomException.NotFound($"User with {createDto.Email} not found.");
             }
 
             // Verify the password
@@ -172,7 +218,14 @@ namespace Backend_Teamwork.src.Services.user
             }
 
             var TokenUtil = new TokenUtils(_configuration);
-            return TokenUtil.GenerateToken(foundUser);
+            var token = TokenUtil.GenerateToken(foundUser);
+
+            if (string.IsNullOrEmpty(token))
+            {
+                throw CustomException.UnAuthorized("Failed to generate token.");
+            }
+
+            return token;
         }
     }
 }
